Split GitHub App callback onto its own endpoint
Route OAuth2 and GitHub App completions through separate HTTP paths
(/connectors/complete vs /connectors/github-app/complete) instead of
sniffing state token types on a shared callback URL.
