+++ .env.example
@@ -148,7 +148,7 @@
 # PROBOD_CONNECTOR_GITHUB_CLIENT_ID=
 # PROBOD_CONNECTOR_GITHUB_CLIENT_SECRET=
 # Configure the GitHub App setup URL as:
-# ${PROBOD_BASE_URL}/api/console/v1/connectors/complete
+# ${PROBOD_BASE_URL}/api/console/v1/connectors/github-app/complete
 # PROBOD_CONNECTOR_GITHUB_APP_ID=
 # PROBOD_CONNECTOR_GITHUB_APP_SLUG=
 # PROBOD_CONNECTOR_GITHUB_APP_PRIVATE_KEY=
